refactor: extract config keybind helpers
